requirements.txt
pycryptodomex==3.11.0
pillow==10.0.0
prompt-toolkit==3.0.39
psutil==5.9.4
